ランキングデータがある時のみゴミ箱ボタンを有効にするため、function書き換え、 修正する。(if文の判定がまだできていない。)

+++ app/src/main/java/com/yui/koalassimonsaysgame_android/rankingPage/RankingPresenter.kt
@@ -1,10 +1,12 @@
 package com.yui.koalassimonsaysgame_android.rankingPage
 
+import android.app.AlertDialog
+import android.content.Context
+import android.content.DialogInterface
 import android.widget.ImageButton
+import com.yui.koalassimonsaysgame_android.ApplicationController
 import com.yui.koalassimonsaysgame_android.Model.UserRankingModel
 import com.yui.koalassimonsaysgame_android.Model.UserRankingModelContract
-import com.yui.koalassimonsaysgame_android.Model.UserRankingModelMock
-import com.yui.koalassimonsaysgame_android.resultPage.ResultActivity
 
 class RankingPresenter(
     private val view: RankingContract.View
@@ -12,20 +14,36 @@ class RankingPresenter(
 
     private val userRankingModel: UserRankingModelContract  = UserRankingModel()
 
+    var context: Context = ApplicationController.applicationContext()
+
     override fun didCreate() {
         val rankingData = userRankingModel.selectData()
         view.setRankingData(rankingData)
     }
 
-    override fun didTapDeleteButton() {
-        userRankingModel.deleteData()
-    }
-
-    override fun disableDeleteButton(button: ImageButton) {
+    override fun checkRankingData(button: ImageButton) {
         val rankingData = userRankingModel.selectData()
 
-        if ( rankingData == null) {
+        if (rankingData.isEmpty()) {
+            // ランキングデータがなかったら、ボタンを無効にする。
             button.isClickable = false
+
+        } else {
+            button.setOnClickListener() {
+
+                val dialog = AlertDialog.Builder(context)
+                dialog.setTitle("ランキングデータを\n削除します。")
+                dialog.setMessage("データは全て削除されます。")
+
+                dialog.setPositiveButton("OK", DialogInterface.OnClickListener { dialog, which ->
+
+                    // OKボタン押したときの処理(ランキングデータを削除する)
+                    userRankingModel.deleteData()
+                })
+
+                dialog.setNegativeButton("キャンセル", null)
+                dialog.show()
+            }
         }
     }
 }
